+++ packages/terra-application/src/application-loading-overlay/ApplicationLoadingOverlayProvider.jsx
@@ -22,9 +22,16 @@ const propTypes = {
    * ApplicationLoadingOverlayProvider.
    */
   scrollRefCallback: PropTypes.func,
+  /**
+   * @private
+   * Prop to not render scroll in  Powerchart.
+   */
+  noScroll: PropTypes.bool,
 };
 
-const ApplicationLoadingOverlayProvider = ({ children, scrollRefCallback, ...customProps }) => {
+const ApplicationLoadingOverlayProvider = ({
+  children, scrollRefCallback, noScroll, ...customProps
+}) => {
   const [registeredLoadingOverlays, setRegisteredLoadingOverlays] = React.useState({});
 
   const contextValue = useMemo(() => ({
@@ -69,6 +76,7 @@ const ApplicationLoadingOverlayProvider = ({ children, scrollRefCallback, ...cus
   );
 
   let className = cx('container');
+  const scrollClass = cx({ 'remove-scroll': noScroll });
   if (customProps.className) {
     className = [className, customProps.className].join(' ');
   }
@@ -79,7 +87,7 @@ const ApplicationLoadingOverlayProvider = ({ children, scrollRefCallback, ...cus
       className={className}
       overlay={overlay}
     >
-      <Scroll refCallback={scrollRefCallback}>
+      <Scroll className={scrollClass} refCallback={scrollRefCallback}>
         <ApplicationLoadingOverlayContext.Provider value={contextValue}>
           {children}
         </ApplicationLoadingOverlayContext.Provider>